feat(flagged): fetch all buckets from the API
